requirements.txt
speechrecognition>=3.8.1
pyttsx3>=2.90
wikipedia>=1.4.0
wolframalpha>=5.0.0
requests>=2.26.0
openai>=0.27.0
python-dotenv>=0.19.0
flask>=2.0.0